LayoutTests/fast/gradients/css3-gradient-parsing-expected.txt

-webkit-linear-gradient


PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black 0%", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(top)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(10deg)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(top center, black 0%)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(10px 20%, black 0%)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(left left)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(left center)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(left 10deg, black 0%)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black, white)", "background-image") is "-webkit-linear-gradient(top, black, white)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black 0, white)", "background-image") is "-webkit-linear-gradient(top, black 0px, white)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black 0%, white)", "background-image") is "-webkit-linear-gradient(top, black 0%, white)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black 0%, white)", "background-image") is "-webkit-linear-gradient(top, black 0%, white)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black, white)", "background-image") is "-webkit-linear-gradient(top, black, white)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(black 10px, white 20px)", "background-image") is "-webkit-linear-gradient(top, black 10px, white 20px)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(top left, black 0%, white)", "background-image") is "-webkit-linear-gradient(left top, black 0%, white)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(top, black 0%, white)", "background-image") is "-webkit-linear-gradient(top, black 0%, white)"
PASS testGradientSpecifiedValue("background-image: -webkit-linear-gradient(10deg, black 0%, white)", "background-image") is "-webkit-linear-gradient(10deg, black 0%, white)"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black 0%", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(top)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(10deg)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(top center, black 0%)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(10px 20%, black 0%)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(left left)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(left center)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(left 10deg, black 0%)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black, white)", "background-image") is "-webkit-linear-gradient(top, rgb(0, 0, 0), rgb(255, 255, 255))"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black 0, white)", "background-image") is "-webkit-linear-gradient(top, rgb(0, 0, 0) 0px, rgb(255, 255, 255))"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black 0%, white)", "background-image") is "-webkit-linear-gradient(top, rgb(0, 0, 0) 0%, rgb(255, 255, 255))"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black 0%, white)", "background-image") is "-webkit-linear-gradient(top, rgb(0, 0, 0) 0%, rgb(255, 255, 255))"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black, white)", "background-image") is "-webkit-linear-gradient(top, rgb(0, 0, 0), rgb(255, 255, 255))"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(black 10px, white 20px)", "background-image") is "-webkit-linear-gradient(top, rgb(0, 0, 0) 10px, rgb(255, 255, 255) 20px)"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(top left, black 0%, white)", "background-image") is "-webkit-linear-gradient(left top, rgb(0, 0, 0) 0%, rgb(255, 255, 255))"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(top, black 0%, white)", "background-image") is "-webkit-linear-gradient(top, rgb(0, 0, 0) 0%, rgb(255, 255, 255))"
PASS testGradientComputedValue("background-image: -webkit-linear-gradient(10deg, black 0%, white)", "background-image") is "-webkit-linear-gradient(10deg, rgb(0, 0, 0) 0%, rgb(255, 255, 255))"
-webkit-repeating-linear-gradient


PASS testGradientSpecifiedValue("background-image: -webkit-repeating-linear-gradient(black, white)", "background-image") is "-webkit-repeating-linear-gradient(top, black, white)"
PASS testGradientComputedValue("background-image: -webkit-repeating-linear-gradient(black, white)", "background-image") is "-webkit-repeating-linear-gradient(top, rgb(0, 0, 0), rgb(255, 255, 255))"
-webkit-radial-gradient


PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(white, black)", "background-image") is "-webkit-radial-gradient(center, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(right bottom, white, black)", "background-image") is "-webkit-radial-gradient(right bottom, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(cover, white, black)", "background-image") is "-webkit-radial-gradient(center, ellipse cover, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(circle, white, black)", "background-image") is "-webkit-radial-gradient(center, circle cover, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(circle contain, white, black)", "background-image") is "-webkit-radial-gradient(center, circle contain, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(top, circle contain, white, black)", "background-image") is "-webkit-radial-gradient(center top, circle contain, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(left top, circle contain, white, black)", "background-image") is "-webkit-radial-gradient(left top, circle contain, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(10px 20%, circle contain, white, black)", "background-image") is "-webkit-radial-gradient(10px 20%, circle contain, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(10px, 20%, circle contain, white, black)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(circle 10px 20%, circle contain, white, black)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(circle 10px, circle contain, white, black)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(center, 10px, white, black)", "background-image") is ""
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(center, 10px 10px, white, black)", "background-image") is "-webkit-radial-gradient(center center, 10px 10px, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(ellipse farthest-corner, white, black)", "background-image") is "-webkit-radial-gradient(center, ellipse farthest-corner, white, black)"
PASS testGradientSpecifiedValue("background-image: -webkit-radial-gradient(circle closest-side, white, black)", "background-image") is "-webkit-radial-gradient(center, circle closest-side, white, black)"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(white, black)", "background-image") is "-webkit-radial-gradient(center, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(right bottom, white, black)", "background-image") is "-webkit-radial-gradient(right bottom, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(cover, white, black)", "background-image") is "-webkit-radial-gradient(center, ellipse cover, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(circle, white, black)", "background-image") is "-webkit-radial-gradient(center, circle cover, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(circle contain, white, black)", "background-image") is "-webkit-radial-gradient(center, circle contain, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(top, circle contain, white, black)", "background-image") is "-webkit-radial-gradient(center top, circle contain, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(left top, circle contain, white, black)", "background-image") is "-webkit-radial-gradient(left top, circle contain, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(10px 20%, circle contain, white, black)", "background-image") is "-webkit-radial-gradient(10px 20%, circle contain, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(10px, 20%, circle contain, white, black)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(circle 10px 20%, circle contain, white, black)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(circle 10px, circle contain, white, black)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(center, 10px, white, black)", "background-image") is "none"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(center, 10px 10px, white, black)", "background-image") is "-webkit-radial-gradient(center center, 10px 10px, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(ellipse farthest-corner, white, black)", "background-image") is "-webkit-radial-gradient(center, ellipse farthest-corner, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS testGradientComputedValue("background-image: -webkit-radial-gradient(circle closest-side, white, black)", "background-image") is "-webkit-radial-gradient(center, circle closest-side, rgb(255, 255, 255), rgb(0, 0, 0))"
-webkit-repeating-radial-gradient


PASS testGradientSpecifiedValue("background-image: -webkit-repeating-radial-gradient(white, black)", "background-image") is "-webkit-repeating-radial-gradient(center, white, black)"
PASS testGradientComputedValue("background-image: -webkit-repeating-radial-gradient(white, black)", "background-image") is "-webkit-repeating-radial-gradient(center, rgb(255, 255, 255), rgb(0, 0, 0))"
PASS successfullyParsed is true

TEST COMPLETE
